Rename the module name.
