+++ docker/config/application.yml
@@ -66,7 +66,7 @@ storage:
 #    url: ${SW_STORAGE_H2_URL:jdbc:h2:mem:skywalking-oap-db}
 #    user: ${SW_STORAGE_H2_USER:sa}
 #  mysql:
-receiver-share-server:
+receiver-sharing-server:
   default:
 receiver-register:
   default:
