+++ oap-server/server-receiver-plugin/skywalking-jvm-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/jvm/provider/JVMModuleProvider.java
@@ -23,7 +23,7 @@ import org.apache.skywalking.oap.server.core.server.GRPCHandlerRegister;
 import org.apache.skywalking.oap.server.library.module.*;
 import org.apache.skywalking.oap.server.receiver.jvm.module.JVMModule;
 import org.apache.skywalking.oap.server.receiver.jvm.provider.handler.*;
-import org.apache.skywalking.oap.server.receiver.share.server.ShareServerModule;
+import org.apache.skywalking.oap.server.receiver.sharing.server.SharingServerModule;
 
 /**
  * @author peng-yongsheng
@@ -46,7 +46,7 @@ public class JVMModuleProvider extends ModuleProvider {
     }
 
     @Override public void start() {
-        GRPCHandlerRegister grpcHandlerRegister = getManager().find(ShareServerModule.NAME).provider().getService(GRPCHandlerRegister.class);
+        GRPCHandlerRegister grpcHandlerRegister = getManager().find(SharingServerModule.NAME).provider().getService(GRPCHandlerRegister.class);
         grpcHandlerRegister.addHandler(new JVMMetricsServiceHandler(getManager()));
         grpcHandlerRegister.addHandler(new JVMMetricReportServiceHandler(getManager()));
     }
@@ -56,6 +56,6 @@ public class JVMModuleProvider extends ModuleProvider {
     }
 
     @Override public String[] requiredModules() {
-        return new String[] {CoreModule.NAME, ShareServerModule.NAME};
+        return new String[] {CoreModule.NAME, SharingServerModule.NAME};
     }
 }
