+++ oap-server/server-receiver-plugin/skywalking-sharing-server-plugin/src/main/resources/META-INF/services/org.apache.skywalking.oap.server.library.module.ModuleDefine
@@ -16,4 +16,4 @@
 #
 #
 
-org.apache.skywalking.oap.server.receiver.share.server.ShareServerModuleProvider
+org.apache.skywalking.oap.server.receiver.sharing.server.SharingServerModule
