+++ oap-server/server-starter/src/main/assembly/application.yml
@@ -66,7 +66,7 @@ storage:
 #    flushInterval: ${SW_STORAGE_ES_FLUSH_INTERVAL:10} # flush the bulk every 10 seconds whatever the number of requests
 #    concurrentRequests: ${SW_STORAGE_ES_CONCURRENT_REQUESTS:2} # the number of concurrent requests
 #  mysql:
-receiver-share-server:
+receiver-sharing-server:
   default:
 receiver-register:
   default:
